Set database to auto-increment book id
Spring/Hibernate returns 'hibernate_sequence' error. This is because it
tries to access hibernate_sequence to get a new id, but this table does
not exist. Instead of using Hibernate to increment id, we will let the
database do the auto-incrementing.

To check: how to use Hibernate to do auto-increment?

https://stackoverflow.com/questions/32968527/hibernate-sequence-doesnt-exist